Committing legacy code from around summer of 2014, with some tweaks
It can only get better from here... The directory structure is not
an exact copy as I made a few structural changes to make it fit for
a Git repository but the code is awful, I'm aware of that. The point
is to make it into something reasonable, and to track the progress.
I've also switched to PostgreSQL as opposed to what DB server the
project originally ran on (MySQL).